requirements.txt
aiogram==3.4.1
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0 